Update visualise imatge tif.py
Codi més optimitzat reduint els bits del float a la meitat.

+++ Homework/visualise imatge tif.py
@@ -11,10 +11,10 @@ nir_path = script_dir / "nir.tif"
 
 # 1. Open files
 with rasterio.open(green_path) as src_green:
-    green = src_green.read(1).astype(float)
+    green = src_green.read(1).astype(np.float32)
 
 with rasterio.open(nir_path) as src_nir:
-    nir = src_nir.read(1).astype(float)
+    nir = src_nir.read(1).astype(np.float32)
 
 # 2. NDWI = (Green - NIR) / (Green + NIR) with error control
 with np.errstate(divide="ignore", invalid="ignore"): # Doesn't show a error each time it divides by zero
